requirements.txt
langgraph 
langchain
huggingface_hub
transformers
torch 
torchvision 
opencv-python 
pillow 
gradio 
numpy
pydantic
langchain
langgraph
langchain-community
langchain-core
python-dotenv
langchain-groq
arxiv
wikipedia
langchain-huggingface